+ Moved AuthViewModel to auth/ directory
