+++ app/src/main/kotlin/io/github/skincanorg/skincan/ui/preference/ProfileActivity.kt
@@ -17,7 +17,7 @@ import by.kirich1409.viewbindingdelegate.viewBinding
 import dagger.hilt.android.AndroidEntryPoint
 import io.github.skincanorg.skincan.databinding.ActivityProfileBinding
 import io.github.skincanorg.skincan.ui.OnboardingActivity
-import io.github.skincanorg.skincan.ui.common.AuthViewModel
+import io.github.skincanorg.skincan.ui.auth.AuthViewModel
 
 @AndroidEntryPoint
 class ProfileActivity : AppCompatActivity() {
